+++ index.js
@@ -1,15 +1,8 @@
 // eslint-disable-next-line max-classes-per-file
-import "./modules/dom-element.js";
-import "./modules/class.js"
-import "./modules/display-books.js";
-import "./modules/add-delete.js";
-import "./modules/display-spa.js";
-
-
-
-
+import './modules/dom-element.js';
+import './modules/book.js';
+import './modules/display-books.js';
+import './modules/add-delete.js';
+import './modules/display-spa.js';
 
 document.getElementById('date').innerHTML = Date();
-
-
-
