+++ modules/book.js
@@ -1,0 +1,7 @@
+/* eslint-disable import/prefer-default-export */
+export class Book {
+  constructor(title, author) {
+    this.title = title;
+    this.author = author;
+  }
+}
